- added up and down functionality for the ladders - created level 1 and level 2
NOTE: There are still a couple of bugs that need to be solve. The ladder's colliders need to be edited a bit more and take up over the platforms a little bit. check out this video for reference (https://www.youtube.com/watch?v=MeBibq_DZ38&index=41&list=PLX-uZVK_0K_6VXcSajfFbXDXndb6AdBLO). Level 2 still needs to be thoroughly edited. The slide functionality needs to be added and the ladder colliders need to be edited.